(bbl-git-lib) script-main: commit_changes: workdir_status_check: improve: more precise workdir check and commit improve
Now the check_workdir function puts in an array
all the changed files found in the workdir, in any state
to provide debug information.

The commit_changes function now checks first the bdm
changed files.
If check passed, the commit only will be performed
when no batch mode is used, or when auto-commit flag
is supplied